+++ docs/release_notes.md
@@ -1,5 +1,11 @@
 # Version history
 
+## v1.2.2
+
+- Some bugs in ACM and MSM are fixed.
+- Update model to `gpt-3.5-turbo`.
+- Removed support for Markdown response because the escape chars are hard to handle. Maybe it will be back in the future.
+
 ## v1.2.1
 
 - A critical bug is fixed.